restructuring for v2.0
