+++ flixkit/flix_service.go
@@ -1,9 +1,9 @@
-package flixkitv2
+package flixkit
 
 import (
 	"context"
 
-	internal "github.com/onflow/flixkit-go/internal/flixkitv2"
+	"github.com/onflow/flixkit-go/internal"
 )
 
 type FlixService interface {
@@ -11,11 +11,13 @@ type FlixService interface {
 	GetTemplate(ctx context.Context, templateName string) (string, error)
 	// GetAndReplaceImports returns the raw flix template with cadence imports replaced
 	GetTemplateAndReplaceImports(ctx context.Context, templateName string, network string) (*FlowInteractionTemplateExecution, error)
+	// GenerateBinding returns the generated binding given the language
+	GetTemplateAndCreateBinding(ctx context.Context, templateName string, lang string) (string, error)
 }
 
 type FlowInteractionTemplateExecution = internal.FlowInteractionTemplateExecution
 
-type FlixServiceConfig = internal.FlixServiceConfig 
+type FlixServiceConfig = internal.FlixServiceConfig
 
 func NewFlixService(config *FlixServiceConfig) FlixService {
 	return internal.NewFlixService(config)
