--- flixkit/testdata/TestJSGenMinScript.golden
@@ -1,28 +1,0 @@
-`/**
-    This binding file was auto generated based on FLIX template v1.0.0.
-    Changes to this file might get overwritten.
-    Note fcl version 1.3.0 or higher is required to use templates.
-**/
-
-import * as fcl from "@onflow/fcl"
-import flixTemplate from "./min.template.json"
-
-/**
-*
-* @param {Object} Parameters - parameters for the cadence
-* @param {Array<string>} Parameters.numbers - : Int
-*/
-export async function request({numbers}) {
-  const info = await fcl.query({
-    template: flixTemplate,
-    args: (arg, t) => [arg(numbers, t.Array(t.Int))]
-  });
-
-  return info
-}
-
-
-
-
-
-`
