--- flixkit/testdata/TestJSGenNoParamsScript.golden
@@ -1,27 +1,0 @@
-`/**
-    This binding file was auto generated based on FLIX template v1.0.0.
-    Changes to this file might get overwritten.
-    Note fcl version 1.3.0 or higher is required to use templates.
-**/
-
-import * as fcl from "@onflow/fcl"
-import flixTemplate from "./min.template.json"
-
-/**
-*
-* No parameters needed.
-*/
-export async function request() {
-  const info = await fcl.query({
-    template: flixTemplate,
-
-  });
-
-  return info
-}
-
-
-
-
-
-`
