--- flixkit/testdata/TestJSGenScript.golden
@@ -1,29 +1,0 @@
-`/**
-    This binding file was auto generated based on FLIX template v1.0.0.
-    Changes to this file might get overwritten.
-    Note fcl version 1.3.0 or higher is required to use templates.
-**/
-
-import * as fcl from "@onflow/fcl"
-import flixTemplate from "./multiply_two_integers.template.json"
-
-/**
-* Multiply two numbers to another
-* @param {Object} Parameters - parameters for the cadence
-* @param {string} Parameters.x - number to be multiplied: Int
-* @param {string} Parameters.y - second number to be multiplied: Int
-*/
-export async function multiplyTwoIntegers({x, y}) {
-  const info = await fcl.query({
-    template: flixTemplate,
-    args: (arg, t) => [arg(x, t.Int), arg(y, t.Int)]
-  });
-
-  return info
-}
-
-
-
-
-
-`
