--- flixkit/testdata/TestJSGenTransaction.golden
@@ -1,29 +1,0 @@
-`/**
-    This binding file was auto generated based on FLIX template v1.0.0.
-    Changes to this file might get overwritten.
-    Note fcl version 1.3.0 or higher is required to use templates.
-**/
-
-import * as fcl from "@onflow/fcl"
-import flixTemplate from "./transfer_token.json"
-
-/**
-* Transfer tokens from one account to another
-* @param {Object} Parameters - parameters for the cadence
-* @param {string} Parameters.amount - The amount of FLOW tokens to send: UFix64
-* @param {string} Parameters.to - The Flow account the tokens will go to: Address
-* @returns {Promise<string>} - returns a promise which resolves to the transaction id
-*/
-export async function transferTokens({amount, to}) {
-  const transactionId = await fcl.mutate({
-    template: flixTemplate,
-    args: (arg, t) => [arg(amount, t.UFix64), arg(to, t.Address)]
-  });
-
-  return transactionId
-}
-
-
-
-
-`
