--- internal/flixkitv2/v1_1/testdata/TestEmptyPragmaWithParameter.golden
@@ -1,23 +1,0 @@
-`{
-    "f_type": "",
-    "f_version": "",
-    "id": "",
-    "data": {
-        "type": "",
-        "interface": "",
-        "messages": null,
-        "cadence": {
-            "body": "",
-            "network_pins": null
-        },
-        "dependencies": null,
-        "parameters": [
-            {
-                "label": "greeting",
-                "index": 0,
-                "type": "String",
-                "messages": []
-            }
-        ]
-    }
-}`
