--- internal/flixkitv2/v1_1/testdata/TestParsePragma/Empty.golden
@@ -1,16 +1,0 @@
-`{
-    "f_type": "",
-    "f_version": "",
-    "id": "",
-    "data": {
-        "type": "",
-        "interface": "",
-        "messages": null,
-        "cadence": {
-            "body": "",
-            "network_pins": null
-        },
-        "dependencies": null,
-        "parameters": null
-    }
-}`
