--- internal/flixkitv2/v1_1/testdata/TestParsePragma/Minimum.golden
@@ -1,16 +1,0 @@
-`{
-    "f_type": "",
-    "f_version": "1.1.0",
-    "id": "",
-    "data": {
-        "type": "",
-        "interface": "",
-        "messages": null,
-        "cadence": {
-            "body": "",
-            "network_pins": null
-        },
-        "dependencies": null,
-        "parameters": null
-    }
-}`
