--- internal/flixkitv2/v1_1/testdata/TestParsePragma/WithParameters.golden
@@ -1,86 +1,0 @@
-`{
-    "f_type": "",
-    "f_version": "1.1.0",
-    "id": "",
-    "data": {
-        "type": "",
-        "interface": "",
-        "messages": [
-            {
-                "key": "title",
-                "i18n": [
-                    {
-                        "tag": "en-US",
-                        "translation": "Update Greeting"
-                    }
-                ]
-            },
-            {
-                "key": "description",
-                "i18n": [
-                    {
-                        "tag": "en-US",
-                        "translation": "Update the greeting on the HelloWorld contract"
-                    }
-                ]
-            }
-        ],
-        "cadence": {
-            "body": "",
-            "network_pins": null
-        },
-        "dependencies": null,
-        "parameters": [
-            {
-                "label": "greeting",
-                "index": 0,
-                "type": "",
-                "messages": [
-                    {
-                        "key": "title",
-                        "i18n": [
-                            {
-                                "tag": "en-US",
-                                "translation": "Greeting"
-                            }
-                        ]
-                    },
-                    {
-                        "key": "description",
-                        "i18n": [
-                            {
-                                "tag": "en-US",
-                                "translation": "The greeting to set on the HelloWorld contract"
-                            }
-                        ]
-                    }
-                ]
-            },
-            {
-                "label": "amount",
-                "index": 1,
-                "type": "",
-                "messages": [
-                    {
-                        "key": "title",
-                        "i18n": [
-                            {
-                                "tag": "en-US",
-                                "translation": "Amount"
-                            }
-                        ]
-                    },
-                    {
-                        "key": "description",
-                        "i18n": [
-                            {
-                                "tag": "en-US",
-                                "translation": "The amount parameter to Test"
-                            }
-                        ]
-                    }
-                ]
-            }
-        ]
-    }
-}`
