--- internal/flixkitv2/v1_1/testdata/TestParsePragma/WithoutParameters.golden
@@ -1,35 +1,0 @@
-`{
-    "f_type": "",
-    "f_version": "1.1.0",
-    "id": "",
-    "data": {
-        "type": "",
-        "interface": "",
-        "messages": [
-            {
-                "key": "title",
-                "i18n": [
-                    {
-                        "tag": "en-US",
-                        "translation": "Update Greeting"
-                    }
-                ]
-            },
-            {
-                "key": "description",
-                "i18n": [
-                    {
-                        "tag": "en-US",
-                        "translation": "Update the greeting on the HelloWorld contract"
-                    }
-                ]
-            }
-        ],
-        "cadence": {
-            "body": "",
-            "network_pins": null
-        },
-        "dependencies": null,
-        "parameters": null
-    }
-}`
