+++ internal/v1_1/testdata/TestParsePragma/Minimum.golden
