+++ internal/v1_1/testdata/TestTransactionValue.golden
